Generate customizable passwords based on a predetermined list of words.
This password generator allows users to select a theme and generate a specified number of passwords based on a list of keywords related to that theme. The keywords are read in from a file, and users can select from a list of themes including animals, cars, Egyptian mythology, Greek/Roman mythology, Star Wars, technology, and Tolkien.

The passwords are generated by selecting random keywords from the list, then adding random special characters and numbers to meet the desired length. The generated passwords are printed to the console.

This version of the password generator includes a welcome screen and a menu allowing users to generate a new password or quit. It also uses the crypto/rand package to generate a cryptographically secure seed for the math/rand package, improving the security of the generated passwords.

Author: Ryan Scrim
Date: 28 Feb 2023
Version: 1.0